% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindVariableFeatures.R
\name{FindVariableFeatures}
\alias{FindVariableFeatures}
\title{Identify variable features by variance decomposition in VarScope object}
\usage{
FindVariableFeatures(
  object,
  assay = "norm",
  var_within_cutoff = 0.4,
  var_between_cutoff = 0.5,
  seed = 1115
)
}
\arguments{
\item{object}{A \code{VarScope} object
containing normalized feature abundance data .}

\item{assay}{Character.
Name of the assay to use for variable feature detection.}

\item{var_within_cutoff}{Numeric.
the quantile threshold for within-group variance.}

\item{var_between_cutoff}{Numeric.
Lower threshold for between-group variance.}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
A \code{VarScope} object containing the variable feature data,
}
\description{
Identify perturbation-driven variable features by decomposing feature-wise
variance into within-group and between-group components. Features showing
low within-group variability and high between-group variability are retained
as variable features.
}
\examples{
NULL
}
