% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.VarScope}
\alias{print.VarScope}
\title{Print VarScope Object}
\usage{
\method{print}{VarScope}(x, ...)
}
\arguments{
\item{x}{VarScope Object}

\item{...}{}
}
\description{
Print VarScope Object
}
\examples{
NULL
}
