% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVarScopeObject.R
\name{CreateVarScopeObject}
\alias{CreateVarScopeObject}
\title{Create a VarScope object}
\usage{
CreateVarScopeObject(
  counts,
  meta.data,
  project = NULL,
  min.group = 0,
  min.features = 0,
  ...
)
}
\arguments{
\item{counts}{Character
The path to a feature table}

\item{meta.data}{Character}

\item{project}{Character}

\item{min.group}{Numeric}

\item{min.features}{Numeric}
}
\value{
A \code{VarScope} object containing raw feature counts, sample metadata,
}
\description{
Create a \code{VarScope} object from an ASV/feature table and corresponding sample metadata.
This function initializes the core container used throughout the VarScopeR workflow.
}
\examples{
\dontrun{
vsobj <- CreateVarScopeObject(
  counts = "inst/extdata/otutab.txt",
  meta.data = "inst/extdata/group.csv",
  project = "ExampleProject"
)

vsobj
}
}
