% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeData.R
\name{NormalizeData}
\alias{NormalizeData}
\title{Normalize feature abundance data in VarScope object}
\usage{
NormalizeData(
  object,
  assay = "raw",
  new.assay = "norm",
  normalization.method = c("none", "log2", "log10", "ln", "rrarefy", "rrarefy_relative",
    "clr"),
  seed = 1115
)
}
\arguments{
\item{object}{A \code{VarScope} object
containing raw feature abundance data .}

\item{assay}{Character.
Name of the input assay to be normalized}

\item{new.assay}{Character.
Name of the normalized data}

\item{normalization.method}{Character.
Data transformation methods applied before correlation analysis.
Options include:
"none" (raw data),
"scale" (z-score standardization),
"center" (mean centering only),
"log2" (log2 transfrom),
"log10" (log10 transfrom),
"ln" (natural transfrom ),
"rrarefy" (random rarefaction using \code{vegan::rrarefy}),
"rrarefy_relative" (rarefy then convert to relative abundance).}

\item{seed}{Integer (default = 1115).
Random seed for reproducibility.}
}
\value{
A \code{VarScope} object containing the normalized data,
}
\description{
Normalize raw feature abundance data (e.g. OTU/ASV tables) stored in a
\code{VarScope} object. The normalized data will be saved as a new assay
within the object, while the original data remain unchanged.
}
\examples{
NULL
}
